requirements.txt
# Requires Python 3.13 or higher
selenium==4.7.1
webdriver-manager==3.8.6
pytest==7.3.1
Appium-Python-Client==2.7.0
